requirements.txt
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
geoalchemy2==0.14.2
psycopg2-binary==2.9.9
pytz==2023.3 